refactor: evaluate archive files in another package
would be used in #2911
